Fix Gatsby build: extract contact form into separate component
Gatsby's static GraphQL query extractor can't parse page files that use
hooks (useState) with block-body arrow functions. Move all form logic
and styled components into a dedicated ContactForm component so the page
file stays as a simple implicit-return arrow function.
